spring5-recipe-app/src/main/java/com/manos/spring5recipeapp/services/RecipeServiceImpl.java
package com.manos.spring5recipeapp.services;

import com.manos.spring5recipeapp.Repositories.RecipeRepository;
import com.manos.spring5recipeapp.models.Recipe;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.Optional;

@Slf4j
@Service
public class RecipeServiceImpl implements RecipeService {

    @Autowired
    RecipeRepository recipeRepository;

    @Override
    public ArrayList<Recipe> getAllRecipes() {

        log.debug("i am in the service");
        return (ArrayList<Recipe>) recipeRepository.findAll();
    }

    @Override
    public Recipe findById(Long id) {
        Optional<Recipe> recipe = recipeRepository.findById(id);
        if(!recipe.isPresent()){
            throw  new RuntimeException("recipe not found");
        }
        return recipe.get();
    }
}
